Silencing GCC compiler warnings

+++ invoke.c
@@ -22,6 +22,7 @@
 #include "listing.h"
 #include "myassert.h"
 #include "label.h"
+#include "fixup.h"
 #include "hll.h"
 
 #include "segment.h"
@@ -3145,7 +3146,7 @@ static int ParamIsString(char *pStr, int param, struct dsym* proc) {
 			if (p->sym.target_type)
 			{
 				type = p->sym.target_type;
-				while (type->target_type && (int)type->target_type > 0x2000)
+				while (type->target_type && (unsigned long)type->target_type > 0x2000)
 				{
 					type = type->target_type;
 					if (type->mem_type == MT_PTR)
@@ -3192,8 +3193,8 @@ static int ParamIsString(char *pStr, int param, struct dsym* proc) {
 static unsigned int hashpjw(const char *s)
 /******************************************/
 {
-	uint_64 fnv_basis = 14695981039346656037;
-	uint_64 register fnv_prime = 1099511628211;
+	uint_64 fnv_basis = 14695981039346656037u;
+	uint_64 register fnv_prime = 1099511628211u;
 	uint_64 h;
 	for (h = fnv_basis; *s; ++s) {
 		h ^= (*s | ' ');
